Fix operator overloading, string/int edge cases; document behaviors
Implementation:
- Comparison operator overloading (lt/gt/lte/gte) now dispatches directly
  to the matching struct method, interpreted by truthiness
- Overloaded `in` (contains) coerces its return value to a boolean
- Creating a string longer than u16::MAX returns an error instead of panicking
- Integer literals are lexed/parsed as i64 and saturated to i32 at compile
  time, so i32::MIN (-2147483648) can now be written; integer negation/abs
  saturate
- print() honors `sep` and `end` keyword arguments and rejects unknown kwargs
- Bare block statements are supported (no trailing semicolon, own scope)

Documentation (LANGUAGE.md):
- Modulo described as the always-non-negative Euclidean remainder
- Negative string indexing, the 65535-byte string limit, the restricted
  escape set (no \r/\0/\x), NaN/inf/-inf truthiness, int() saturation of
  non-finite floats, empty-list pop() returning None, and cyclic-printing
  with `...`

Added regression tests for the new behaviors.

+++ shimlang/src/lex.rs
@@ -48,7 +48,11 @@ pub enum Token {
     LCurly,
     RCurly,
     None,
-    Integer(i32),
+    // Integer literals are lexed as i64 so that a magnitude one past
+    // i32::MAX (2147483648) can be carried through to the parser, where a
+    // leading minus folds it into i32::MIN. Bare out-of-range literals are
+    // saturated to the i32 range at compile time.
+    Integer(i64),
     Float(f32),
     Bool(bool),
     Identifier(Vec<u8>),
